Split reducer into own file
